requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arviz==0.12.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
cffi==1.15.1
cftime==1.6.2
charset-normalizer==2.1.1
clikit==0.6.2
crashtest==0.3.1
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
fastjsonschema==2.16.2
fonttools==4.38.0
frozenlist==1.3.3
httpstan==4.6.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.5
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
MarkupSafe==2.1.2
marshmallow==3.19.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.4
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netCDF4==1.6.2
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.21.6
packaging==23.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.14.0
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.19.3
pysimdjson==3.2.0
pystan==3.3.0
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
scipy==1.7.3
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
terminado==0.17.1
tinycss2==1.2.1
torch==1.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.4.0
wcwidth==0.2.6
webargs==8.2.0
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.5
xarray==0.20.2
xarray-einstats==0.2.2
yarl==1.8.2
zipp==3.12.1
